voteItemController実装
投票の新規作成、更新、削除

+++ src/main/java/com/example/app/controller/VoteItemController.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import com.example.app.domain.VoteItem;
 import com.example.app.mapper.VoteItemMapper;
@@ -22,10 +23,7 @@ public class VoteItemController {
 	
 	//Homeページへ
 	@GetMapping("/home")
-	public String homePage(
-			Model model,
-			HttpServletRequest request
-			) {
+	public String homePage(Model model,HttpServletRequest request) {
 		HttpSession session = request.getSession();
 	  String userName = (String) session.getAttribute("user");
 	  
@@ -36,7 +34,7 @@ public class VoteItemController {
 	  
 	  
 	  //ほかのユーザーの投票の投稿見る、投票する、投票結果見る
-	  //全ての投票を取得・
+	  //全ての投票を取得（他のユーザーの投稿を表示）
 	  model.addAttribute("voteItems", voteItemmapper.selectAll());
 	  
 	  //新規投票作成用フォーム
@@ -48,26 +46,22 @@ public class VoteItemController {
 	//投票の新規作成・変更更新・削除
 	@PostMapping("/home")
 	public String createVoteItem(
-			VoteItem voteItem
+			@RequestParam("action")String action,
+			VoteItem voteItem,
+			@RequestParam(value= "id", required = false)Integer id
 			) {
-		//新規作成
-		voteItemmapper.addVoteItem(voteItem);
+		if("create".equals(action)) {
+			//新規作成
+			voteItemmapper.addVoteItem(voteItem);			
+		}else if("update".equals(action)) {
+		     // 更新
+			voteItem.setVoteItemId(id);//対象のIDセット
+			voteItemmapper.updateVoteItem(voteItem);
+		} else if("delete".equals(action) && id != null)  {
+			//削除
+			voteItemmapper.deleteVoteItem(id);
+		}
 		return "redirect:/home";
 	}
 
-	@PostMapping("/home")
-	public String deleteVoteItem(Integer id) {
-		//削除
-		voteItemmapper.deleteVoteItem(id);
-		return "redirect:/home";
-	}
-	
-	@PostMapping("/home")
-	public String updateVoteItem(VoteItem voteItem) {
-		//更新
-		voteItemmapper.updateVoteItem(voteItem);
-		return "redirect:/home";
-	}
-	
-	
 }
